refactor: extract powershell and notifications into top-level modules
Move PowerShellExecutor and ps_quote/ps_quote_for_xml/check_pid_exists/
run_with_graceful_timeout out of desktop/ into a new windows_mcp.powershell
package. Move Desktop.send_notification into windows_mcp.notifications. None
of this logic depends on Desktop state. Update callers (registry, tools/shell,
tools/notification, tests/test_registry) to use the new paths and drop the
now-unused logging import from desktop/utils.py.
