+++ src/windows_mcp/notifications/__init__.py
@@ -1,0 +1,5 @@
+from windows_mcp.notifications.service import send_notification
+
+__all__ = [
+    'send_notification',
+]
