+++ src/windows_mcp/powershell/__init__.py
@@ -1,0 +1,5 @@
+from windows_mcp.powershell.service import PowerShellExecutor
+
+__all__ = [
+    'PowerShellExecutor',
+]
