+++ src/windows_mcp/tools/shell.py
@@ -2,7 +2,7 @@
 
 from mcp.types import ToolAnnotations
 from windows_mcp.infrastructure import with_analytics
-from windows_mcp.desktop.powershell import PowerShellExecutor
+from windows_mcp.powershell import PowerShellExecutor
 from fastmcp import Context
